# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API MiniMarket Plus",
        "description": "Microservicio backend de MiniMarket Plus. Expone la gestion de\nproductos, categorias, inventario, carrito de compras, ventas,\ndetalle de ventas y usuarios del sistema.\n\nEsta documentacion se genera con OpenAPI Specification (OAS 3)\nmediante springdoc-openapi, asegurando la consistencia entre el\ncodigo y la documentacion de la API.",
        "contact": {
            "name": "Equipo Backend - MiniMarket Plus",
            "email": "soporte@minimarketplus.cl"
        },
        "license": {
            "name": "Uso academico Duoc UC",
            "url": "https://www.duoc.cl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Entorno de desarrollo local"
        }
    ],
    "security": [
        {
            "autenticacionBasica": []
        }
    ],
...
    "paths": {
...
        "/api/inventario/{id}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener movimiento por ID",
                "description": "Devuelve la informacion detallada de un movimiento de inventario segun su identificador.",
                "operationId": "obtenerMovimientoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del movimiento de inventario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimiento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Movimiento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar movimiento de inventario",
                "description": "Actualiza los datos de un movimiento de inventario existente identificado por su ID.",
                "operationId": "actualizarMovimiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del movimiento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Movimiento actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Movimiento no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar movimiento de inventario",
                "description": "Elimina un movimiento de inventario por su identificador.",
                "operationId": "eliminarMovimiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del movimiento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Movimiento eliminado (sin contenido)"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Movimiento no encontrado"
                    }
                }
            }
        },
...
        "/api/inventario": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Listar movimientos de inventario",
                "description": "Obtiene la lista completa de movimientos de inventario registrados en MiniMarket Plus.",
                "operationId": "listarMovimientosDeInventario",
                "responses": {
                    "200": {
                        "description": "Lista de movimientos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Registrar movimiento de inventario",
                "description": "Registra un nuevo movimiento de inventario (entrada o salida de stock) para un producto existente.",
                "operationId": "registrarMovimiento",
                "requestBody": {
                    "description": "Datos del movimiento de inventario a registrar",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            },
                            "examples": {
                                "Movimiento nuevo": {
                                    "description": "Movimiento nuevo",
                                    "value": {
                                        "producto": {
                                            "id": 1
                                        },
                                        "cantidad": 20,
                                        "tipoMovimiento": "Entrada",
                                        "fechaMovimiento": "2026-07-06T15:30:00.000+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Movimiento registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            }
        },
...
        "/api/ventas/{id}": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Obtener venta por ID",
                "description": "Devuelve la informacion detallada de una venta registrada segun su identificador.",
                "operationId": "obtenerVentaPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venta"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Venta no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Rol": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del rol.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre del rol segun convencion de Spring Security.",
                        "example": "ROLE_ADMIN"
                    }
                },
                "description": "Rol de autorizacion que agrupa permisos de acceso (RBAC)."
            },
            "Usuario": {
                "required": [
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del usuario.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "description": "Nombre de usuario unico para autenticarse.",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "description": "Contrasena del usuario (se almacena cifrada con BCrypt). Solo escritura.",
                        "writeOnly": true,
                        "example": "admin123"
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "description": "Roles asignados al usuario (por ejemplo ROLE_ADMIN).",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    }
                },
                "description": "Usuario del sistema con sus credenciales y roles de acceso."
            },
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico de la categoria.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre unico de la categoria.",
                        "example": "Bebidas"
                    },
                    "productos": {
                        "type": "array",
                        "description": "Productos asociados a esta categoria.",
                        "readOnly": true,
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "description": "Categoria que agrupa productos del catalogo (por ejemplo Bebidas o Abarrotes)."
            },
            "Producto": {
                "required": [
                    "categoria",
                    "nombre",
                    "precio",
                    "stock"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del producto.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre comercial del producto.",
                        "example": "Agua Mineral 1L"
                    },
                    "precio": {
                        "type": "number",
                        "description": "Precio de venta unitario en pesos chilenos.",
                        "format": "double",
                        "example": 890.0
                    },
                    "stock": {
                        "type": "integer",
                        "description": "Cantidad disponible en stock.",
                        "format": "int32",
                        "example": 50
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "description": "Producto disponible en el catalogo de MiniMarket Plus."
            },
            "Inventario": {
                "required": [
                    "cantidad",
                    "fechaMovimiento",
                    "producto",
                    "tipoMovimiento"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del movimiento de inventario.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "description": "Cantidad de unidades del movimiento.",
                        "format": "int32",
                        "example": 20
                    },
                    "tipoMovimiento": {
                        "type": "string",
                        "description": "Tipo de movimiento: Entrada (reabastecimiento) o Salida (consumo).",
                        "example": "Entrada",
                        "enum": [
                            "Entrada",
                            "Salida"
                        ]
                    },
                    "fechaMovimiento": {
                        "type": "string",
                        "description": "Fecha y hora en que se registro el movimiento.",
                        "format": "date-time",
                        "example": "2026-07-06T15:30:00Z"
                    }
                },
                "description": "Movimiento de inventario (entrada o salida de stock) asociado a un producto."
            },
            "DetalleVenta": {
                "required": [
                    "cantidad",
                    "precio",
                    "producto",
                    "venta"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del detalle de venta.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "venta": {
                        "$ref": "#/components/schemas/Venta"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "description": "Cantidad de unidades vendidas del producto.",
                        "format": "int32",
                        "example": 2
                    },
                    "precio": {
                        "type": "number",
                        "description": "Precio unitario aplicado al momento de la venta.",
                        "format": "double",
                        "example": 890.0
                    }
                },
                "description": "Linea de detalle de una venta: producto, cantidad y precio unitario aplicado."
            },
            "Venta": {
                "required": [
                    "fecha",
                    "usuario"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico de la venta.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "fecha": {
                        "type": "string",
                        "description": "Fecha y hora de la venta.",
                        "format": "date-time",
                        "example": "2026-07-06T15:45:00Z"
                    },
                    "detalles": {
                        "type": "array",
                        "description": "Detalle de productos, cantidades y precios de la venta.",
                        "items": {
                            "$ref": "#/components/schemas/DetalleVenta"
                        }
                    }
                },
                "description": "Venta registrada en MiniMarket Plus, con su detalle de productos vendidos."
            },
...
        },
        "securitySchemes": {
            "autenticacionBasica": {
                "type": "http",
                "description": "Autenticacion HTTP Basic. Use el usuario y la contrasena semilla (admin / admin123) en el boton Authorize para consumir los endpoints protegidos.",
                "name": "autenticacionBasica",
                "scheme": "basic"
            }
        }
    }
}